Move files to more appropriate folders
- Use src/ instead of lib/
- Move the main index.jsx into src/
- Move the CSS to where it is used
- Do not use a folder components/
- Move test to what is tested
