+++ package.json
@@ -22,8 +22,8 @@
     "dev": "nrfconnect-scripts build-watch",
     "webpack": "nrfconnect-scripts build-dev",
     "build": "nrfconnect-scripts build-prod",
-    "lint": "nrfconnect-scripts lint lib index.jsx",
-    "lintfix": "nrfconnect-scripts lint --fix lib index.jsx",
+    "lint": "nrfconnect-scripts lint src",
+    "lintfix": "nrfconnect-scripts lint --fix src",
     "test": "nrfconnect-scripts test",
     "test-watch": "nrfconnect-scripts test --watch",
     "clean": "npm run clean-dist && npm run clean-modules",
